feat: Schedule, Theme의 생성/삭제 요청 URI를 admin/ 하위로 이동
- testMethod 수정 작업 필요한 상태
