+++ src/main/java/nextstep/admin/AdminController.java
@@ -1,15 +1,45 @@
 package nextstep.admin;
 
+import lombok.RequiredArgsConstructor;
+import nextstep.schedule.ScheduleRequest;
+import nextstep.schedule.ScheduleService;
+import nextstep.theme.ThemeRequest;
+import nextstep.theme.ThemeService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
 
 @Controller
 @RequestMapping("/admin")
+@RequiredArgsConstructor
 public class AdminController {
-    @GetMapping
-    public ResponseEntity<Void> admin() {
-        return ResponseEntity.ok().build();
+    private final ScheduleService scheduleService;
+    private final ThemeService themeService;
+
+    @PostMapping("/schedules")
+    public ResponseEntity<Void> createSchedule(@RequestBody ScheduleRequest scheduleRequest) {
+        Long id = scheduleService.create(scheduleRequest);
+        return ResponseEntity.created(URI.create("/schedules/" + id)).build();
+    }
+
+    @DeleteMapping("/schedules/{id}")
+    public ResponseEntity<Void> deleteSchedule(@PathVariable Long id) {
+        scheduleService.deleteById(id);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/themes")
+    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
+        Long id = themeService.create(themeRequest);
+        return ResponseEntity.created(URI.create("/themes/" + id)).build();
+    }
+
+    @DeleteMapping("/themes/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
