+++ src/main/java/nextstep/schedule/ScheduleController.java
@@ -1,9 +1,11 @@
 package nextstep.schedule;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -16,23 +18,10 @@ public class ScheduleController {
         this.scheduleService = scheduleService;
     }
 
-    @PostMapping
-    public ResponseEntity<Void> createSchedule(@RequestBody ScheduleRequest scheduleRequest) {
-        Long id = scheduleService.create(scheduleRequest);
-        return ResponseEntity.created(URI.create("/schedules/" + id)).build();
-    }
-
     @GetMapping
-    public ResponseEntity<List<ScheduleResponse>> showReservations(@RequestParam Long themeId, @RequestParam String date) {
+    public ResponseEntity<List<ScheduleResponse>> showSchedules(@RequestParam Long themeId, @RequestParam String date) {
         List<Schedule> schedules = scheduleService.findByThemeIdAndDate(themeId, date);
         List<ScheduleResponse> scheduleResponses = schedules.stream().map(ScheduleResponse::fromEntity).collect(Collectors.toList());
         return ResponseEntity.ok().body(scheduleResponses);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        scheduleService.deleteById(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
