+++ src/main/java/nextstep/theme/ThemeController.java
@@ -1,9 +1,10 @@
 package nextstep.theme;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -16,24 +17,11 @@ public class ThemeController {
         this.themeService = themeService;
     }
 
-    @PostMapping
-    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
-        Long id = themeService.create(themeRequest);
-        return ResponseEntity.created(URI.create("/themes/" + id)).build();
-    }
-
     @GetMapping
     public ResponseEntity<List<ThemeResponse>> showThemes() {
         List<Theme> themes = themeService.findAll();
         List<ThemeResponse> themeResponses = themes.stream().map(ThemeResponse::fromEntity).collect(Collectors.toList());
 
         return ResponseEntity.ok().body(themeResponses);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.delete(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
